+++ rmk/CHANGELOG.md
@@ -60,7 +60,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Support read default keymap from `keyboard.toml`, see https://haobogu.github.io/rmk/configuration.html#keymap-config
+- Support read default keymap from `keyboard.toml`, see https://haobogu.github.io/rmk/keyboard_configuration.html#keymap-config
 
 ## [0.1.17] - 2024-06-04